pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = ["yang-lin <casongg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.65.1"
grpcio-tools = "^1.65.1"
fastapi = "^0.111.1"
pydantic-settings = "^2.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
